tiledb_group_dump_str should takes a uint_8 (#3019)
* tiledb_group_dump_str should takes a uint_8

This was merged as taking an int but is hould be a uint_8 to align with
other bools.

* Fix index bound check for group member access

* Handle file:// vs plain path for removal

* [ci] Pin werkzeug to fix GCS emulator startup

+++ tiledb/sm/c_api/tiledb_experimental.h
@@ -830,8 +830,10 @@ TILEDB_EXPORT int32_t tiledb_group_get_query_type(
  * @return  `TILEDB_OK` for success and `TILEDB_ERR` for error.
  */
 TILEDB_EXPORT int32_t tiledb_group_dump_str(
-    tiledb_ctx_t* ctx, tiledb_group_t* group, char** dump_ascii, int recursive)
-    TILEDB_NOEXCEPT;
+    tiledb_ctx_t* ctx,
+    tiledb_group_t* group,
+    char** dump_ascii,
+    uint8_t recursive) TILEDB_NOEXCEPT;
 
 #ifdef __cplusplus
 }
